workers rate and add worker problem :)

+++ app/src/main/java/com/example/abdulaziz/myapplication/viewworkerinfo.java
@@ -64,7 +64,7 @@ public class viewworkerinfo extends AppCompatActivity {
                 WorkerSkills.setText(w1.getSkills());
                 WorkerPrice.setText(String.valueOf(w1.getPrice()));
                 WorkerIncome.setText(String.valueOf(w1.getTotalIncome()));
-                rate.setRating(w1.getTotalRate()/w1.getnRate());
+                rate.setRating((float) (w1.getTotalRate()/w1.getnRate()));
                 Picasso.get().load(w1.getPicture()).into(imageView);
             }
